Implement LCD in specific file
